spl_token_2022::extension::confidential_transfer

Module ciphertext_extraction

Ciphertext extraction and proof related helper logic

This submodule should be removed with the next upgrade to the Solana program Ciphertext extraction and proof related helper logic

This submodule should be removed with the next upgrade to the Solana program

Structs

SourceDecryptHandles

The ElGamal ciphertext decryption handle pertaining to the low and high bits of the transfer amount under the source public key of the transfer.

TransferProofContextInfo

The proof context information needed to process a [Transfer] instruction.

TransferPubkeysInfo

The transfer public keys associated with a transfer.

TransferWithFeeProofContextInfo

The proof context information needed to process a [Transfer] instruction with fee.

TransferWithFeePubkeysInfo

The transfer public keys associated with a transfer with fee.

Functions

transfer_amount_source_ciphertext

Extract the transfer amount ciphertext encrypted under the source ElGamal public key.