spl_token_2022::extension::confidential_transfer

Module instruction

Confidential Transfer Extension instructions

Modules

batched_grouped_ciphertext_validity

The batched grouped-ciphertext validity proof instruction.

batched_range_proof

The batched range proof instructions.

ciphertext_ciphertext_equality

The ciphertext-ciphertext equality proof instruction.

ciphertext_commitment_equality

The ciphertext-commitment equality proof instruction.

errors

fee_sigma

The fee sigma proof instruction.

grouped_ciphertext_validity

The grouped-ciphertext validity proof instruction.

pubkey_validity

The public-key validity proof instruction.

range_proof

The range proof instruction.

transfer

withdraw

zero_balance

The zero-balance proof instruction.

Structs

ApplyPendingBalanceData

Data expected by ConfidentialTransferInstruction::ApplyPendingBalance

BatchedGroupedCiphertext2HandlesValidityProofContext

BatchedGroupedCiphertext2HandlesValidityProofData

The instruction data that is needed for the ProofInstruction::VerifyBatchedGroupedCiphertextValidity instruction.

BatchedRangeProofContext

The context data needed to verify a range-proof for a Pedersen committed value.

BatchedRangeProofU64Data

The instruction data that is needed for the ProofInstruction::VerifyBatchedRangeProofU64 instruction.

BatchedRangeProofU128Data

The instruction data that is needed for the ProofInstruction::VerifyBatchedRangeProofU128 instruction.

BatchedRangeProofU256Data

The instruction data that is needed for the ProofInstruction::BatchedRangeProofU256Data instruction.

CiphertextCiphertextEqualityProofContext

The context data needed to verify a ciphertext-ciphertext equality proof.

CiphertextCiphertextEqualityProofData

The instruction data that is needed for the ProofInstruction::VerifyCiphertextCiphertextEquality instruction.

CiphertextCommitmentEqualityProofContext

The context data needed to verify a ciphertext-commitment equality proof.

CiphertextCommitmentEqualityProofData

The instruction data that is needed for the ProofInstruction::VerifyCiphertextCommitmentEquality instruction.

CloseSplitContextStateAccounts

Accounts needed if close_split_context_state_on_execution flag is enabled on a transfer.

ConfigureAccountInstructionData

Data expected by ConfidentialTransferInstruction::ConfigureAccount

ContextStateInfo

Pubkeys associated with a context state account to be used as parameters to functions.

DepositInstructionData

Data expected by ConfidentialTransferInstruction::Deposit

EmptyAccountInstructionData

Data expected by ConfidentialTransferInstruction::EmptyAccount

FeeParameters

FeeSigmaProofContext

The context data needed to verify a pubkey validity proof.

FeeSigmaProofData

The instruction data that is needed for the ProofInstruction::VerifyFeeSigma instruction.

GroupedCiphertext2HandlesValidityProofContext

GroupedCiphertext2HandlesValidityProofData

The instruction data that is needed for the ProofInstruction::VerifyGroupedCiphertextValidity instruction.

InitializeMintData

Data expected by ConfidentialTransferInstruction::InitializeMint

ProofContextState

The proof context account state

PubkeyValidityData

The instruction data that is needed for the ProofInstruction::VerifyPubkeyValidity instruction.

PubkeyValidityProofContext

The context data needed to verify a pubkey validity proof.

RangeProofContext

The context data needed to verify a range-proof for a committed value in a Pedersen commitment.

RangeProofU64Data

The instruction data that is needed for the ProofInstruction::VerifyRangeProofU64 instruction.

TransferData

The instruction data that is needed for the ProofInstruction::VerifyTransfer instruction.

TransferInstructionData

Data expected by ConfidentialTransferInstruction::Transfer

TransferProofContext

The context data needed to verify a transfer proof.

TransferSplitContextStateAccounts

Type for split transfer (without fee) instruction proof context state account addresses intended to be used as parameters to functions.

TransferWithFeeData

The instruction data that is needed for the ProofInstruction::TransferWithFee instruction.

TransferWithFeeProofContext

The context data needed to verify a transfer-with-fee proof.

TransferWithFeeSplitContextStateAccounts

Type for split transfer (with fee) instruction proof context state account addresses intended to be used as parameters to functions.

TransferWithSplitProofsInstructionData

Data expected by ConfidentialTransferInstruction::TransferWithSplitProofs

UpdateMintData

Data expected by ConfidentialTransferInstruction::UpdateMint

WithdrawData

The instruction data that is needed for the ProofInstruction::VerifyWithdraw instruction.

WithdrawInstructionData

Data expected by ConfidentialTransferInstruction::Withdraw

WithdrawProofContext

The context data needed to verify a withdraw proof.

ZeroBalanceProofContext

The context data needed to verify a zero-balance proof.

ZeroBalanceProofData

The instruction data that is needed for the ProofInstruction::ZeroBalance instruction.

Enums

ConfidentialTransferInstruction

Confidential Transfer extension instructions

ProofInstruction

ProofType

Traits

Pod

Marker trait for “plain old data”.

ZkProofData

Functions

apply_pending_balance

Create a ApplyPendingBalance instruction

approve_account

Create an ApproveAccount instruction

close_context_state

Create a CloseContextState instruction.

configure_account

Create a ConfigureAccount instruction

deposit

Create a Deposit instruction

disable_confidential_credits

Create a DisableConfidentialCredits instruction

disable_non_confidential_credits

Create a DisableNonConfidentialCredits instruction

empty_account

Create a EmptyAccount instruction

enable_confidential_credits

Create a EnableConfidentialCredits instruction

enable_non_confidential_credits

Create a EnableNonConfidentialCredits instruction

initialize_mint

Create a InitializeMint instruction

inner_apply_pending_balance

Create a inner ApplyPendingBalance instruction

inner_configure_account

Create a ConfigureAccount instruction

inner_empty_account

Create an inner EmptyAccount instruction

inner_transfer

Create a inner Transfer instruction

inner_transfer_with_fee

Create a inner Transfer instruction with fee

inner_withdraw

Create a inner Withdraw instruction

transfer

Create a Transfer instruction with regular (no-fee) proof

transfer_with_fee

Create a Transfer instruction with fee proof

transfer_with_fee_and_split_proofs

Create a TransferWithSplitProof instruction with fee

transfer_with_split_proofs

Create a TransferWithSplitProof instruction without fee

update_mint

Create a UpdateMint instruction

verify_batched_verify_range_proof_u64

Create a VerifyBatchedRangeProofU64 instruction.

verify_batched_verify_range_proof_u128

Create a VerifyBatchedRangeProofU128 instruction.

verify_batched_verify_range_proof_u256

Create a VerifyBatchedRangeProofU256 instruction.

verify_ciphertext_ciphertext_equality

Create a VerifyCiphertextCiphertextEquality instruction.

verify_ciphertext_commitment_equality

Create a VerifyCiphertextCommitmentEquality instruction.

verify_pubkey_validity

Create a VerifyPubkeyValidity instruction.

verify_range_proof_u64

Create a VerifyRangeProofU64 instruction.

verify_transfer

Create a VerifyTransfer instruction.

verify_transfer_with_fee

Create a VerifyTransferWithFee instruction.

verify_withdraw

Create a VerifyWithdraw instruction.

verify_zero_balance

Create a VerifyZeroBalance instruction.

withdraw

Create a Withdraw instruction

Derive Macros

Pod

Derive the Pod trait for a struct