spl_token_2022::extension::confidential_transfer::instruction
Module batched_grouped_ciphertext_validity

Structs

BatchedGroupedCiphertext2HandlesValidityProofContext

BatchedGroupedCiphertext2HandlesValidityProofData

The instruction data that is needed for the ProofInstruction::VerifyBatchedGroupedCiphertextValidity instruction.

BatchedGroupedCiphertext3HandlesValidityProofContext

BatchedGroupedCiphertext3HandlesValidityProofData

The instruction data that is needed for the ProofInstruction::VerifyBatchedGroupedCiphertext3HandlesValidity instruction.