spl_token_2022::extension::confidential_transfer::instruction
Module percentage_with_cap

The percentage-with-cap proof instruction.

The percentage-with-cap proof is defined with respect to three Pedersen commitments that encodes values referred to as a percentage, delta, and claimed amounts. The proof certifies that either

the percentage amount is equal to a constant (referred to as the max_value)

the delta and claimed amounts are equal

Structs

PercentageWithCapProofContext

The context data needed to verify a percentage-with-cap proof.

PercentageWithCapProofData

The instruction data that is needed for the ProofInstruction::VerifyPercentageWithCap instruction.