spl_token_2022::extension
Module confidential_transfer_fee

Confidential Transfer Fee extension

Modules

account_info

Confidential Transfer Fee extension account information needed for instructions

instruction

Confidential transfer fee extension instructions

processor

Confidential transfer fee extension processor

Structs

ConfidentialTransferFeeAmount

Confidential transfer fee

ConfidentialTransferFeeConfig

Confidential transfer fee extension data for mints

Type Aliases

EncryptedFee

ElGamal ciphertext containing a transfer fee

EncryptedWithheldAmount

ElGamal ciphertext containing a withheld fee in an account