spl_token_2022::extension

Module transfer_hook

Transfer Hook extension

Modules

instruction

Instructions for the TransferHook extension

processor

Instruction processor for the TransferHook extension

Structs

TransferHook

Transfer hook extension data for mints.

TransferHookAccount

Indicates that the tokens from this account belong to a mint with a transfer hook

Functions

get_program_id

Attempts to get the transfer hook program id from the TLV data, returning None if the extension is not found

set_transferring

Helper function to set the transferring flag before calling into transfer hook

unset_transferring

Helper function to unset the transferring flag after a transfer