spl_token_2022

Module proof

Helper for processing instruction data from ZK Token proof program

Structs

SplitContextStateAccountsConfig

Instruction options for when using split context state accounts

Enums

ProofLocation

A proof location type meant to be used for arguments to instruction constructors.

Functions

decode_proof_instruction_context

Decodes the proof context data associated with a zero-knowledge proof instruction.