spl_token_2022::extension::confidential_transfer
Module instruction

Confidential Transfer Extension instructions

Modules

batched_grouped_ciphertext_validity

batched_range_proof

The batched range proof instructions.

ciphertext_ciphertext_equality

The ciphertext-ciphertext equality proof instruction.

ciphertext_commitment_equality

The ciphertext-commitment equality proof instruction.

errors

grouped_ciphertext_validity

percentage_with_cap

The percentage-with-cap proof instruction.

pod

pubkey_validity

The public-key validity proof instruction.

zero_ciphertext

The zero-ciphertext proof instruction.

Structs

ApplyPendingBalanceData

Data expected by ConfidentialTransferInstruction::ApplyPendingBalance

BatchedGroupedCiphertext2HandlesValidityProofContext

BatchedGroupedCiphertext2HandlesValidityProofData

The instruction data that is needed for the ProofInstruction::VerifyBatchedGroupedCiphertextValidity instruction.

BatchedGroupedCiphertext3HandlesValidityProofContext

BatchedGroupedCiphertext3HandlesValidityProofData

The instruction data that is needed for the ProofInstruction::VerifyBatchedGroupedCiphertext3HandlesValidity instruction.

BatchedRangeProofContext

The context data needed to verify a range-proof for a Pedersen committed value.

BatchedRangeProofU64Data

The instruction data that is needed for the ProofInstruction::VerifyBatchedRangeProofU64 instruction.

BatchedRangeProofU128Data

The instruction data that is needed for the ProofInstruction::VerifyBatchedRangeProofU128 instruction.

BatchedRangeProofU256Data

The instruction data that is needed for the ProofInstruction::BatchedRangeProofU256Data instruction.

CiphertextCiphertextEqualityProofContext

The context data needed to verify a ciphertext-ciphertext equality proof.

CiphertextCiphertextEqualityProofData

The instruction data that is needed for the ProofInstruction::VerifyCiphertextCiphertextEquality instruction.

CiphertextCommitmentEqualityProofContext

The context data needed to verify a ciphertext-commitment equality proof.

CiphertextCommitmentEqualityProofData

The instruction data that is needed for the ProofInstruction::VerifyCiphertextCommitmentEquality instruction.

ConfigureAccountInstructionData

Data expected by ConfidentialTransferInstruction::ConfigureAccount

DepositInstructionData

Data expected by ConfidentialTransferInstruction::Deposit

EmptyAccountInstructionData

Data expected by ConfidentialTransferInstruction::EmptyAccount

GroupedCiphertext2HandlesValidityProofContext

GroupedCiphertext2HandlesValidityProofData

The instruction data that is needed for the ProofInstruction::VerifyGroupedCiphertextValidity instruction.

GroupedCiphertext3HandlesValidityProofContext

GroupedCiphertext3HandlesValidityProofData

The instruction data that is needed for the ProofInstruction::VerifyGroupedCiphertext3HandlesValidity instruction.

InitializeMintData

Data expected by ConfidentialTransferInstruction::InitializeMint

PercentageWithCapProofContext

The context data needed to verify a percentage-with-cap proof.

PercentageWithCapProofData

The instruction data that is needed for the ProofInstruction::VerifyPercentageWithCap instruction.

ProofContextState

The proof context account state

PubkeyValidityProofContext

The context data needed to verify a pubkey validity proof.

PubkeyValidityProofData

The instruction data that is needed for the ProofInstruction::VerifyPubkeyValidity instruction.

TransferInstructionData

Data expected by ConfidentialTransferInstruction::Transfer

TransferWithFeeInstructionData

Data expected by ConfidentialTransferInstruction::TransferWithFee

UpdateMintData

Data expected by ConfidentialTransferInstruction::UpdateMint

WithdrawInstructionData

Data expected by ConfidentialTransferInstruction::Withdraw

ZeroCiphertextProofContext

The context data needed to verify a zero-ciphertext proof.

ZeroCiphertextProofData

The instruction data that is needed for the ProofInstruction::ZeroCiphertext instruction.

Enums

ConfidentialTransferInstruction

Confidential Transfer extension instructions

ProofInstruction

ProofType

Traits

ZkProofData

Functions

apply_pending_balance

Create a ApplyPendingBalance instruction

approve_account

Create an ApproveAccount instruction

configure_account

Create a ConfigureAccount instruction

configure_account_with_registry

Create a ConfigureAccountWithRegistry instruction

deposit

Create a Deposit instruction

disable_confidential_credits

Create a DisableConfidentialCredits instruction

disable_non_confidential_credits

Create a DisableNonConfidentialCredits instruction

empty_account

Create a EmptyAccount instruction

enable_confidential_credits

Create a EnableConfidentialCredits instruction

enable_non_confidential_credits

Create a EnableNonConfidentialCredits instruction

initialize_mint

Create a InitializeMint instruction

inner_apply_pending_balance

Create a inner ApplyPendingBalance instruction

inner_configure_account

Create a ConfigureAccount instruction

inner_empty_account

Create an inner EmptyAccount instruction

inner_transfer

Create an inner Transfer instruction

inner_transfer_with_fee

Create an inner TransferWithFee instruction

inner_withdraw

Create a inner Withdraw instruction

transfer

Create a Transfer instruction

transfer_with_fee

Create a TransferWithFee instruction

update_mint

Create a UpdateMint instruction

withdraw

Create a Withdraw instruction